requirements.txt
psycopg2==2.9.3
matplotlib==3.6.0
tabulate==0.6.10
python-dotenv==0.21.0